Refactor policy core modules
